refactor: extract compare modal components
